+++ CryptocurrencyExchange/Models/FutureDto.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Symbol { get; set; } = string.Empty;
-        public double Margin { get; set; }
+        public decimal Margin { get; set; }
         public double EntryPrice { get; set; }
         public int Leverage { get; set; }
         public PositionType Position { get; set; }
