+++ CryptocurrencyExchange/Services/Interfaces/IWalletDomainService.cs
@@ -4,7 +4,7 @@ namespace CryptocurrencyExchange.Services.Interfaces
 {
     public interface IWalletDomainService
     {
-        void Buy(WalletItem usdt, WalletItem coin, decimal usd, double coinPrice);
+        void Buy(WalletItem usdt, WalletItem coin, decimal usd, decimal coinPrice);
         void Sell(WalletItem usdt, WalletItem coin, decimal amount, decimal coinPrice);
     }
 }
